refactor: 등록하기 페이지에서 BoardsWrite 컴포넌트 분리 및 재사용
